various osx fixes
