+++ server/src/libgen/libgenThread.cpp
@@ -7,6 +7,39 @@
 #include <errno.h>
 #include <sys/time.h>
 
+// missing phread_mutex_timedlock
+//https://lists.freedesktop.org/archives/spice-devel/2010-September/001231.html
+#ifdef __APPLE__
+#include <pthread.h>
+#include <errno.h>
+int pthread_mutex_timedlock(pthread_mutex_t *mutex, const struct timespec
+*abs_timeout)
+{
+    int pthread_rc;
+    struct timespec remaining, slept, ts;
+
+    remaining = *abs_timeout;
+    while ((pthread_rc = pthread_mutex_trylock(mutex)) == EBUSY) {
+        ts.tv_sec = 0;
+        ts.tv_nsec = (remaining.tv_sec > 0 ? 10000000 : min(remaining.tv_nsec,10000000));
+        nanosleep(&ts, &slept);
+        ts.tv_nsec -= slept.tv_nsec;
+        if (ts.tv_nsec <= remaining.tv_nsec) {
+            remaining.tv_nsec -= ts.tv_nsec;
+        }
+        else {
+            remaining.tv_sec--;
+            remaining.tv_nsec = (1000000 - (ts.tv_nsec - remaining.tv_nsec));
+        }
+        if (remaining.tv_sec < 0 || (!remaining.tv_sec && remaining.tv_nsec <= 0)) {
+            return ETIMEDOUT;
+        }
+    }
+
+    return pthread_rc;
+}
+#endif
+
 /**
 * @brief 
 *
